main.py
Addition of main file that when executed records info about the usage of RAM and CPU and eventually, when the user interrupts the program (Ctrl + C), the maximum usage of RAM and CPU is displayed and the info is kept inside an Excel file and a matplotlib graph is created and saved in the same directory main.py is put. At the moment, the program can't handle the transformation into an .exe file, therefore, I must investigate to learn what is preventing the compilation to happen (as a first hypothesis, I believe that the problem is related to the presence of the matplotlib. If that confirms to be true, then I'll have to learn some other way to automize the generation of the graph).